adding favorites to localStorage (completed)
