+++ src/index.js
@@ -14,7 +14,7 @@ import './js/news-data/news-category';
 // MAIN SCRIPTS
 import './js/news-data/news-popular';
 
-import favorite from './js/favorite';
+import favorite from './js/favoriteAddToLS';
 
 mobileMenu();
 headerResponsive();
